Rename directory redis_record --> yarr
