+++ lib/yarr/attribute.rb
