+++ lib/yarr/logger.rb
