+++ lib/yarr/lua_script_reader.rb
