+++ lib/yarr/range_interval.rb
@@ -1,5 +1,5 @@
 # typed: strict
-require 'redis_record/attribute'
+require 'yarr/attribute'
 
 class RedisRecord::RangeInterval < T::Struct
   prop :min, T.nilable(RedisRecord::Attribute::RangeIndexType), default: nil
