+++ lib/yarr/server_scripts.rb
