+++ lib/yarr/server_scripts/create_hash_returning_id.erb.lua
