+++ lib/yarr/server_scripts/delete_hash.erb.lua
