+++ lib/yarr/server_scripts/find_by_attr.erb.lua
