+++ lib/yarr/server_scripts/shared/index_helper_methods.erb.lua
