+++ lib/yarr/server_scripts/shared/lua_helper_methods.erb.lua
