+++ lib/yarr/server_scripts/update_hash.erb.lua
